requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=gpu
absl-py=0.11.0=pyhd3eb1b0_1
argon2-cffi=20.1.0=py37he774522_1
astunparse=1.6.3=py_0
atomicwrites=1.4.0=py_0
attrs=20.2.0=py_0
av=8.0.2=py37h0621dd0_1
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.2.1=py_0
blinker=1.4=py37_0
brotlipy=0.7.0=py37h2bbff1b_1003
bzip2=1.0.8=he774522_0
ca-certificates=2020.12.8=haa95532_0
cached-property=1.5.2=pypi_0
cachetools=4.2.0=pyhd3eb1b0_0
certifi=2020.12.5=py37haa95532_0
cffi=1.14.2=py37h7a1dbc1_0
chardet=3.0.4=py37haa95532_1003
click=7.1.2=py_0
cloudpickle=1.6.0=py_0
cmake=3.18.2=hab937c2_0
colorama=0.4.3=py_0
configparser=5.0.1=pypi_0
cryptography=3.3.1=py37hcd4344a_0
cudatoolkit=10.2.89=h74a9793_1
cycler=0.10.0=py37_0
cytoolz=0.10.1=py37he774522_0
dask-core=2.26.0=py_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
dlib=19.21.1=pypi_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.3=py37_0
ffmpeg=1.4=pypi_0
freeglut=3.2.1=h0e60522_0
freetype=2.10.2=hd328e21_0
gast=0.3.3=py_0
gitdb=4.0.5=pypi_0
gitpython=3.1.11=pypi_0
google-auth=1.21.3=py_0
google-auth-oauthlib=0.4.2=pyhd3eb1b0_2
google-pasta=0.2.0=py_0
grpcio=1.31.0=py37he7da953_0
h5py=3.1.0=pypi_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=67.1=h33f27b4_0
idna=2.10=py_0
imageio=2.9.0=py_0
importlib-metadata=1.7.0=py37_0
importlib_metadata=1.7.0=0
imutils=0.5.3=py37h03978a9_3
iniconfig=1.0.1=py_0
intel-openmp=2020.2=254
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.18.1=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=py_0
jasper=2.0.14=hdc05fd1_1
jedi=0.17.2=py37_0
jinja2=2.11.2=py_0
joblib=0.16.0=py_0
jpeg=9d=h8ffe710_0
jsonschema=3.2.0=py37_1
jupyter_client=6.1.6=py_0
jupyter_core=4.6.3=py37_0
keras=2.4.3=pypi_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_1
kiwisolver=1.2.0=py37h74a9793_0
libblas=3.8.0=20_mkl
libcblas=3.8.0=20_mkl
libclang=10.0.1=default_hf44288c_1
liblapack=3.8.0=20_mkl
liblapacke=3.8.0=20_mkl
libopencv=4.5.0=py37_2
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.13.0.1=h200bbdf_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.1.0=h56a325e_1
libuv=1.39.0=he774522_0
libwebp-base=1.1.0=h8ffe710_3
lz4-c=1.9.2=h62dcd97_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
m2w64-pkg-config=0.29.1=2
markdown=3.3.3=py37haa95532_0
markupsafe=1.1.1=py37hfa6e2cd_1
matplotlib=3.3.1=0
matplotlib-base=3.3.1=py37hba9282a_0
mistune=0.8.4=py37hfa6e2cd_1001
mkl=2020.2=256
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.2.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
more-itertools=8.5.0=py_0
msys2-conda-epoch=20160418=1
mtcnn=0.1.0=pypi_0
nbconvert=5.6.1=py37_1
nbformat=5.0.7=py_0
networkx=2.5=py_0
ninja=1.10.1=py37h7ef1ec2_0
notebook=6.1.1=py37_0
numpy=1.19.1=py37h5510c5b_0
numpy-base=1.19.1=py37ha3acd2a_0
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
opencv=4.5.0=py37_2
opencv-python=4.4.0.42=pypi_0
openssl=1.1.1i=h2bbff1b_0
opt_einsum=3.1.0=py_0
packaging=20.4=py_0
pandas=1.1.5=py37hf11a4ad_0
pandoc=2.10.1=0
pandocfilters=1.4.2=py37_1
parso=0.7.0=py_0
pickleshare=0.7.5=py37_1001
pillow=7.2.0=py37hcc1f983_0
pip=20.2.2=py37_0
pkgconfig=1.4.0=py37_0
pluggy=0.13.1=py37_0
prometheus_client=0.8.0=py_0
promise=2.3=pypi_0
prompt-toolkit=3.0.7=py_0
protobuf=3.13.0.1=py37ha925a31_1
psutil=5.7.3=pypi_0
py=1.9.0=py_0
py-opencv=4.5.0=py37hc6149b9_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pygments=2.7.1=py_0
pyjwt=1.7.1=py37_0
pyopenssl=20.0.0=pyhd3eb1b0_1
pyparsing=2.4.7=py_0
pyqt=5.12.3=py37h03978a9_6
pyqt-impl=5.12.3=py37hf2a7229_6
pyqt5-sip=4.19.18=py37hf2a7229_6
pyqtchart=5.12=py37hf2a7229_6
pyqtwebengine=5.12.1=py37hf2a7229_6
pyreadline=2.1=py37_1
pyrsistent=0.17.3=py37he774522_0
pysocks=1.7.1=py37_1
pytest=6.0.2=py37_0
python=3.7.9=h60c2a47_0
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.6.0=py3.7_cuda102_cudnn7_0
pytz=2020.4=pyhd3eb1b0_0
pywavelets=1.1.1=py37he774522_0
pywin32=227=py37he774522_1
pywinpty=0.5.7=py37_0
pyyaml=5.3.1=py37he774522_1
pyzmq=19.0.2=py37ha925a31_1
qt=5.12.9=hb2cf2c5_0
requests=2.25.0=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.6=py_0
scikit-image=0.16.2=py37h47e9c7a_0
scikit-learn=0.23.2=py37h47e9c7a_0
scipy=1.5.2=py37h9439919_0
seaborn=0.11.1=pyhd3eb1b0_0
send2trash=1.5.0=py37_0
sentry-sdk=0.19.5=pypi_0
setuptools=49.6.0=py37_0
shortuuid=1.0.1=pypi_0
sip=4.19.8=py37h6538335_0
six=1.15.0=py_0
smmap=3.0.4=pypi_0
sqlite=3.33.0=h2a8f88b_0
subprocess32=3.5.4=pypi_0
tensorboard=2.3.0=pyh4dce500_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.3.0=mkl_py37h10aaca4_0
tensorflow-base=2.3.0=eigen_py37h17acbac_0
tensorflow-estimator=2.3.0=pyheb71bc4_0
tensorflow-gpu=2.3.0=he13fc11_0
termcolor=1.1.0=py37_1
terminado=0.8.3=py37_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
toml=0.10.1=py_0
toolz=0.10.0=py_0
torchvision=0.7.0=py37_cu102
tornado=6.0.4=py37he774522_1
tqdm=4.48.2=py_0
traitlets=4.3.3=py37_0
urllib3=1.25.11=py_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wandb=0.10.12=pypi_0
watchdog=1.0.1=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
werkzeug=1.0.1=py_0
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=py37_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37_0
winpty=0.4.3=4
wrapt=1.12.1=py37he774522_1
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
zeromq=4.3.2=ha925a31_3
zipp=3.1.0=py_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
